Refactored Task, Event, Deadline and Todo into their own classes
