+++ src/main/java/ToDo.java
@@ -1,0 +1,10 @@
+public class ToDo extends Task {
+    ToDo(String s) {
+        super(s);
+    }
+
+    @Override
+    public String toString() {
+        return String.format("[T]%s", super.toString());
+    }
+}
